create `CoinBalanceEvent`
