--- mm2src/coins/events.rs
@@ -1,7 +1,0 @@
-use crate::MmCoin;
-use async_trait::async_trait;
-
-#[async_trait]
-pub trait BalanceEvent: MmCoin {
-    async fn stream(&self);
-}
