--- mm2src/coins/tendermint/balance_event.rs
@@ -1,9 +1,0 @@
-use async_trait::async_trait;
-
-use super::TendermintCoin;
-use crate::events::BalanceEvent;
-
-#[async_trait]
-impl BalanceEvent for TendermintCoin {
-    async fn stream(&self) {}
-}
